Put middleware in its own folder
