+++ lib/survey_gizmo/connection.rb
@@ -31,28 +31,12 @@ module SurveyGizmo
         @connection ||= Faraday.new(options) do |connection|
           connection.request :url_encoded
 
-          connection.response :surveygizmo_data
-          connection.response :pester
+          connection.response :parse_survey_gizmo_data
+          connection.response :pester_survey_gizmo
           connection.response :logger, @logger, bodies: true if SurveyGizmo.configuration.api_debug
           connection.response :json, content_type: /\bjson$/
 
           connection.adapter Faraday.default_adapter
-        end
-      end
-    end
-  end
-
-  class PesterMiddleware < Faraday::Middleware
-    Faraday::Response.register_middleware(pester: self)
-
-    def call(environment)
-      Pester.survey_gizmo_ruby.retry do
-        @app.call(environment).on_complete do |response|
-          fail RateLimitExceededError if response.status == 429
-          fail BadResponseError, "Bad response code #{response.status} in #{response.inspect}" unless response.status == 200
-          unless response.body['result_ok'] && response.body['result_ok'].to_s.downcase == 'true'
-            fail BadResponseError, response.inspect unless response_ok?(response)
-          end
         end
       end
     end
